move to a flat components structure, where each component owns its CSS module
