+++ src/main.tsx
@@ -3,7 +3,7 @@ import 'hyperparam/global.css'
 import 'hyperparam/hyperparam.css'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './components/App.js'
+import App from './components/App/App.js'
 
 const app = document.getElementById('app')
 if (!app) throw new Error('No app element')
